Use backup pollser instead of connectivity watcher
