+++ grpc.gemspec
@@ -230,10 +230,9 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/tsi/transport_security_adapter.h )
   s.files += %w( src/core/tsi/transport_security_interface.h )
   s.files += %w( src/core/ext/transport/chttp2/server/chttp2_server.h )
-  s.files += %w( src/core/ext/filters/client_channel/channel_connectivity_internal.h )
+  s.files += %w( src/core/ext/filters/client_channel/backup_poller.h )
   s.files += %w( src/core/ext/filters/client_channel/client_channel.h )
   s.files += %w( src/core/ext/filters/client_channel/client_channel_factory.h )
-  s.files += %w( src/core/ext/filters/client_channel/connectivity_watcher.h )
   s.files += %w( src/core/ext/filters/client_channel/connector.h )
   s.files += %w( src/core/ext/filters/client_channel/http_connect_handshaker.h )
   s.files += %w( src/core/ext/filters/client_channel/http_proxy.h )
@@ -605,12 +604,11 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/tsi/transport_security_adapter.cc )
   s.files += %w( src/core/ext/transport/chttp2/server/chttp2_server.cc )
   s.files += %w( src/core/ext/transport/chttp2/client/secure/secure_channel_create.cc )
+  s.files += %w( src/core/ext/filters/client_channel/backup_poller.cc )
   s.files += %w( src/core/ext/filters/client_channel/channel_connectivity.cc )
-  s.files += %w( src/core/ext/filters/client_channel/channel_connectivity_internal.cc )
   s.files += %w( src/core/ext/filters/client_channel/client_channel.cc )
   s.files += %w( src/core/ext/filters/client_channel/client_channel_factory.cc )
   s.files += %w( src/core/ext/filters/client_channel/client_channel_plugin.cc )
-  s.files += %w( src/core/ext/filters/client_channel/connectivity_watcher.cc )
   s.files += %w( src/core/ext/filters/client_channel/connector.cc )
   s.files += %w( src/core/ext/filters/client_channel/http_connect_handshaker.cc )
   s.files += %w( src/core/ext/filters/client_channel/http_proxy.cc )
