+++ tools/doxygen/Doxyfile.core.internal
@@ -907,16 +907,14 @@ src/core/ext/census/trace_string.h \
 src/core/ext/census/tracing.cc \
 src/core/ext/census/tracing.h \
 src/core/ext/filters/client_channel/README.md \
+src/core/ext/filters/client_channel/backup_poller.cc \
+src/core/ext/filters/client_channel/backup_poller.h \
 src/core/ext/filters/client_channel/channel_connectivity.cc \
-src/core/ext/filters/client_channel/channel_connectivity_internal.cc \
-src/core/ext/filters/client_channel/channel_connectivity_internal.h \
 src/core/ext/filters/client_channel/client_channel.cc \
 src/core/ext/filters/client_channel/client_channel.h \
 src/core/ext/filters/client_channel/client_channel_factory.cc \
 src/core/ext/filters/client_channel/client_channel_factory.h \
 src/core/ext/filters/client_channel/client_channel_plugin.cc \
-src/core/ext/filters/client_channel/connectivity_watcher.cc \
-src/core/ext/filters/client_channel/connectivity_watcher.h \
 src/core/ext/filters/client_channel/connector.cc \
 src/core/ext/filters/client_channel/connector.h \
 src/core/ext/filters/client_channel/http_connect_handshaker.cc \
